server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pymongo==4.7.0
boto3==1.34.162
httpx==0.27.0
pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
pytesseract==0.3.10
openpyxl==3.1.5

# OCR / Vision
easyocr==1.7.2
torch==2.8.0
torchvision==0.23.0
scikit-image==0.25.2

# Data handling
pandas==2.3.3

# local / testing
mongomock==4.1.2
python-multipart==0.0.9

# Twilio official lib
twilio==8.11.0

#Authetication
python-jose==3.3.0
passlib[bcrypt]==1.7.4

openpyxl>=3.1.5
